# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payments Gate",
        "description": "API for handling payment transactions",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/create-transaction": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Create new transaction",
                "description": "Supported payment methods: TPAY, PAYNOW, NODA.",
                "operationId": "0b946754e09859751bb0d713cef1404d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "email",
                                    "name",
                                    "paymentMethod",
                                    "notificationUrl",
                                    "returnUrl"
                                ],
                                "properties": {
                                    "amount": {
                                        "type": "string",
                                        "example": "12.34"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jan.kowalski@gmail.com"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Jan Kowalski"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "PLN"
                                    },
                                    "paymentMethod": {
                                        "type": "string",
                                        "enum": [
                                            "TPAY",
                                            "PAYNOW",
                                            "NODA"
                                        ],
                                        "example": "TPAY"
                                    },
                                    "notificationUrl": {
                                        "type": "string",
                                        "example": "https://test.payment-gate.pl/callback"
                                    },
                                    "returnUrl": {
                                        "type": "string",
                                        "example": "https://test.payment-gate.pl/return-url"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transaction created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "link": {
                                            "type": "string",
                                            "example": "https://link-to-payment"
                                        },
                                        "transactionUuid": {
                                            "type": "string",
                                            "example": "c6e2c816-3f5e-417b-9e91-a794223aa903"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid transaction data provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "The transaction could not be completed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/refund-payment": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Refund payment",
                "operationId": "58f3d9871ca37f979a5dee7bd8c7f5e3",
                "parameters": [
                    {
                        "name": "signature",
                        "in": "header",
                        "description": "HMAC signature of the request body",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "0798f264-f887-4102-b0ce-9d27e3076cc5"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "transactionUuid"
                                ],
                                "properties": {
                                    "transactionUuid": {
                                        "type": "string",
                                        "example": "c6e2c816-3f5e-417b-9e91-a794223aa903"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment refund",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "string",
                                            "example": "Refund"
                                        },
                                        "transactionUuid": {
                                            "type": "string",
                                            "example": "c6e2c816-3f5e-417b-9e91-a794223aa903"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid transaction data provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Refund payment not completed."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/transaction/{uuid}/status": {
            "get": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Get transaction status by UUID",
                "operationId": "ab9deb5f5dc75490b7ca4373978cf3ef",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Transaction UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "c6e2c816-3f5e-417b-9e91-a794223aa903"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "SUCCESS"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 199.99
                                        },
                                        "currency": {
                                            "type": "string",
                                            "example": "PLN"
                                        },
                                        "paymentMethod": {
                                            "type": "string",
                                            "example": "TPAY"
                                        },
                                        "returnUrl": {
                                            "type": "string",
                                            "format": "uri",
                                            "example": "https://test.payment-gate.pl/return-url"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Transaction not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Transaction not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "Wprowadź swój klucz API",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "api_key": []
        }
    ]
}